Added functions for easier testing
Adjusted the model creations to vary the output layer depending on whether it is a classification or regression task. Added a function to create and train all of the specified models: MANN, MANN with dropout, Uniform model for each activation function, and Sequential model in which the order is chosen arbitrarily based on ease of updating gradients (can adjust for better orientation if need to). The function creates and trains all models on the data. The function splits the data into testing and training sets. The required inputs to the function are the normalized feature data (x) and one-hot encoded target variables (y) if task is classification otherwise the target predictions. num_classes is also required as an argument but is not used when used for regression. The function returns a dictionary containing all trained models.